+++ routes/user.js
@@ -1,10 +1,10 @@
-var getUser=require("../middlewares").getUser;
-var getUserAuctions=require("../middlewares").getUserAuctions;
-var renderMW=require("../middlewares").renderMW;
-var updateUser=require("../middlewares").updateUser;
-var getAdmin=require("../middlewares").getAdmin;
-var getAuctionList=require("../middlewares").getAuctionList;
-var getItems=require("../middlewares").getItems;
+var getUser=require("../middlewares/user").getUser;
+var getUserAuctions=require("../middlewares/auctions").getUserAuctions;
+var renderMW=require("../middlewares/generic").renderMW;
+var updateUser=require("../middlewares/user").updateUser;
+var getAdmin=require("../middlewares/generic").getAdmin;
+var getAuctionList=require("../middlewares/auctions").getAuctionList;
+var getItems=require("../middlewares/item").getItems;
 
 module.exports=function(app){
     app.get('/user',
